% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.R
\name{get_tcga_goi}
\alias{get_tcga_goi}
\alias{get_tcga_goi.RangedSummarizedExperiment}
\alias{get_tcga_goi.list}
\alias{get_tcga_goi.character}
\title{Get the expression values out of the TCGA SummarizedExperiment object}
\usage{
get_tcga_goi(data, goi, ...)

\method{get_tcga_goi}{RangedSummarizedExperiment}(data, goi, ...)

\method{get_tcga_goi}{list}(data, goi, element = 1, ...)

\method{get_tcga_goi}{character}(data, goi, ...)
}
\arguments{
\item{data}{A SummarizedExperiment object.}

\item{goi}{A character vector containing the gene symbols of interest.}

\item{...}{Extra arguments}

\item{element}{Index of name of the list element to use.}
}
\description{
A wrapper around \link[SummarizedExperiment]{assay} to extract the TCGA assay, filter out the genes of interest and
always return gene_symbols instead of Ensembl gene IDs.
}
