% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.R
\name{separate_barcode}
\alias{separate_barcode}
\title{Separate TCGA barcode}
\usage{
separate_barcode(data, col = barcode, remove = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{col}{Column name containing the barcode}

\item{remove}{If TRUE, remove the column containing the barcode.}
}
\description{
A wrapper around \link[tidyr]{separate} to split TCGA barcodes into individual variables as described \href{https://docs.gdc.cancer.gov/Encyclopedia/pages/TCGA_Barcode/#reading-barcodes}{here}.
}
